third_party/blink/web_tests/mhtml/page_with_javascript-expected.txt

CONSOLE ERROR: Blocked script execution in 'page_with_javascript.mht' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
